requirements.txt
accelerate==1.5.2
transformers==4.50.1
opencv-python==4.11.0.86
Pillow
torch
runpod==1.6.0
